pyproject.toml
[project]
name = "scratch-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.11.7",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.1.1",
    "watchdog>=6.0.0",
]
